Fixed CSS

+++ build.sh
@@ -5,5 +5,5 @@ npm run build
 cd ..
 rm -rf coco-bot/build
 mv frontend/build coco-bot/build
-rm coco-bot.zip
+if [ -f coco-bot.zip ]; then rm coco-bot.zip; fi
 zip -r coco-bot.zip coco-bot/*
